+++ Changelog.md
@@ -92,6 +92,7 @@ With the port to Bootstrap 3, app/views/terms/default.haml has a new structure. 
 
 ## Bug fixes
 * Fix mention autocomplete when pasting the username [#6510](https://github.com/diaspora/diaspora/pull/6510)
+* Use and update updated\_at for notifications [#6573](https://github.com/diaspora/diaspora/pull/6573)
 
 ## Features